backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
amazon-transcribe
websockets
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
asyncpg
python-jose
passlib
requests
aiofile
pydub
ffmpeg-python
sse-starlette
google-generativeai